requirements.txt
pip>=19.1.1
Scrapy==2.1.0
webdriver_manager==3.1.0
selenium==3.141.0
pandas==1.0.1
scrapyd>=1.2.1
scrapy-redis
logparser>=0.8.2
matplotlib==2.2.5
numpy==1.18.1
pathlib2==2.3.5
scrapy-rotating-proxies==0.6.2
webdriver-manager==3.1.0
psycopg2
python-dotenv
requests
scrapyd-client
twisted
sqlalchemy
pgeocode